+++ packages/components/fields/time-field/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './time-field';
+
+export { default as version } from './version';
