--- src/components/fields/time-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './time-field';
